add: delete properties with empty string values
When the user does not enter any value (e.g. for 'proxy' option),
he is asking to use the value configured up in the config chain
(e.g. in the global npmrc file). This is achieved by not including
such option in the user ini file at all.

`inquirer.prompt` returns an empty string as the answer when the user
does not enter any value and there is no default offered. sl-registry
has to delete such answers to provide the expected behaviour.