promote: use $HOME/.npmrc when appropriate
When one of the --from/--to registries is currently in use,
the promote command uses $HOME/.npmrc instead of the strong-registry
ini file.

That way the current (possibly updated) configuration is used;
the changes made by `npm login` are stored in npmrc and thus
sync-ed to the ini file later.